# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    GraphDatabaseStatus:
      type: object
      properties:
        activeBackend:
          type: string
          description: Currently active graph database backend
        connected:
          type: boolean
          description: Whether the graph database is reachable
        claimCount:
          type: integer
          format: int64
          description: Number of claims in the graph database
        version:
          type: string
          description: Graph database version or implementation info
        rebuildNeeded:
          type: boolean
          description: |
            True when the active graph store has no claims but RDF assets exist in
            storage — typical after a backend switch or fresh start. The admin UI
            surfaces a rebuild prompt that calls `POST /admin/graph/rebuild`. Non-RDF
            assets are excluded from this check because they do not produce graph
            claims even after a successful rebuild.
        rdfAssetCount:
          type: integer
          format: int64
          description: |
            Number of active RDF assets in storage that would be re-indexed by a
            rebuild. Surfaced alongside `rebuildNeeded` so the UI can show how many
            assets the rebuild will process.
    GraphDatabaseSwitchResult:
      type: object
      properties:
        message:
          type: string
          description: Human-readable status message
    GraphStatus:
      type: object
      properties:
        backend:
          type: string
          description: The graph database backend type (NEO4J, FUSEKI, NONE)
        enabled:
          type: boolean
          description: Whether the graph store is enabled
        healthy:
          type: boolean
          description: Whether the graph store backend is healthy
        activeAssetCount:
          type: integer
          format: int64
          description: Number of active assets in PostgreSQL storage
        claimCountInGraph:
          type: integer
          format: int64
          description: Number of claim triples/nodes in the graph database
        assetCountInGraph:
          type: integer
          format: int64
          description: Number of distinct assets in the graph database
        syncAssessment:
          type: string
          description: Sync assessment (empty, in-sync, out-of-sync, unknown, disabled)
...
    RebuildStatus:
      type: object
      properties:
        total:
          type: integer
          format: int64
          description: Total number of assets to process
        processed:
          type: integer
          format: int64
          description: Number of assets processed so far
        percentComplete:
          type: integer
          description: Completion percentage (0-100)
        running:
          type: boolean
          description: Whether a rebuild is currently running
        complete:
          type: boolean
          description: Whether the rebuild has completed
        failed:
          type: boolean
          description: Whether the rebuild has failed
        errorMessage:
          type: string
          description: Error message if the rebuild failed
        errors:
          type: integer
          format: int64
          description: Number of assets that failed during processing
        durationMs:
          type: integer
          format: int64
          description: Duration in milliseconds since rebuild started
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >-
        Forbidden. The user does not have the permission to execute this
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ClientError:
      description: >-
        May contain hints how to solve the error or indicate what was wrong in
        the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /admin/graph-database:
    get:
      tags:
        - AdminGraph
      summary: Get graph database status
      operationId: getGraphDatabaseStatus
      security:
        - jwt: []
      responses:
        '200':
          description: Graph database status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDatabaseStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/graph-database/switch:
    post:
      tags:
        - AdminGraph
      summary: Switch the graph database backend (requires restart)
      operationId: switchGraphDatabase
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                backend:
                  type: string
                  description: Target backend (NEO4J, FUSEKI, NONE)
              required:
                - backend
      responses:
        '200':
          description: Switch request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDatabaseSwitchResult'
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/graph/rebuild:
    post:
      tags:
        - AdminGraph
      summary: Trigger an asynchronous graph rebuild
      operationId: triggerGraphRebuild
      security:
        - jwt: []
      responses:
        '202':
          description: Graph rebuild started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebuildStatus'
        '409':
          description: A graph rebuild is already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebuildStatus'
        '503':
          description: Graph store is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/graph/rebuild/status:
    get:
      tags:
        - AdminGraph
      summary: Get the current graph rebuild status
      operationId: getGraphRebuildStatus
      security:
        - jwt: []
      responses:
        '200':
          description: Current rebuild status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebuildStatus'
  /admin/graph/status:
    get:
      tags:
        - AdminGraph
      summary: Get the current graph store status
      operationId: getGraphStatus
      security:
        - jwt: []
      responses:
        '200':
          description: Graph store status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphStatus'
...